+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Added
 - Operations such as ``-=`` and ``++`` for store variables in C++.
 - YAML export of store meta-data.
 - ``stored::MuxLayer`` to multiplex multiple protocol layers over a single connection.
+- ``stored::Aes256Layer`` for encrypted communication using AES-256 CTR.
 
 Changed
 ```````